Update README.md
The benefits for an individual developer is he/she can look at different version because they are saved as snapshots, also can work on different computers and/or off-line. If something just created turns out not to be right, one can revert to an earlier version and startd again from there. 
When working as a team on a software development project, version control systems such as Git allow every team member to work on their own copy, experiment, submit the changes they have made to a repository and those change are then pulled from there to be merged with the main project after ensuring there are no conflicts with the changes made by other team members.
In addition, since each developer has a copy of the whole thing, they can, after making some changes, test it as a whole to see if the code works. Having a history of the changes, or 'patches', helps when debugging a complex code and needing to go back and look at historic changes.